tg-intel/apps/backend/requirements.txt
fastapi[standard]==0.115.12
uvicorn[standard]==0.34.3
pydantic>=2.7.0,<3.0.0
python-dotenv==1.0.1
telethon==1.35.0
APScheduler==3.10.4
psycopg[binary,pool]==3.2.10
httpx==0.27.0


